modules/swcore: assure proper pages storage to linked list when LL goes through S_SOF_ON_WR
If LL FSM goes through S_SOF_ON_WR state we have to ensure that the state of
mpm_dlast and new_pck_first_page is properly stored for S_IDLE state.
Otherwise, if RCV FSM is in S_RCV_DATA only for one cycle (and then goes to
DROP) we were not storing the start page with _valid_ and _eof_ high.